+++ lib/test_prof/autopilot/dsl.rb
@@ -16,6 +16,10 @@ require "test_prof/autopilot/stack_prof/printer"
 require "test_prof/autopilot/stack_prof/writer"
 require "test_prof/autopilot/stack_prof/profiling_executor"
 
+require "test_prof/autopilot/factory_default/printer"
+require "test_prof/autopilot/factory_default/writer"
+require "test_prof/autopilot/factory_default/profiling_executor"
+
 module TestProf
   module Autopilot
     # Module contains all available DSL instructions