+++ lib/test_prof/autopilot/factory_default/printer.rb
@@ -2,18 +2,20 @@
 
 module TestProf
   module Autopilot
-    module FactoryProf
+    module FactoryDefault
       # Module is used for printing :factory_prof report
       module Printer
-        Registry.register(:factory_prof_printer, self)
+        Registry.register(:factory_default_prof_printer, self)
 
         class PrinterError < StandardError; end
 
         def print_report(report)
           result = report.result
 
-          # TODO: Move total_time to the report
-          TestProf::FactoryProf::Printers::Simple.dump(result, start_time: TestProf.now, threshold: 0)
+          require "test_prof/factory_default"
+          profiler = TestProf::FactoryDefault::Profiler.new
+          profiler.data.merge!(result)
+          profiler.print_report
         end
 
         module_function :print_report