+++ spec/test_prof/autopilot/event_prof/printer_spec.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "spec_helper"
+
 describe TestProf::Autopilot::EventProf::Printer do
   subject { described_class }