+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add FactoryDefault profiler support. ([@palkan][])
+
 ## 0.1.0 (2023-11-21)
 
 - Initial public release. ([@palkan][])